+++ R/report_fishery_mortality.R
@@ -131,10 +131,10 @@ plot_stock_mortality <- function(fram_db, run_id, stock_id,
   }
 
   if(!is.null(filters_list) && !is.list(filters_list)){
-    cli::cli_abort("If provided, filters_list must be a list of fishery filter functions.")
+    fram_abort("If provided, filters_list must be a list of fishery filter functions.")
   }
   if(!is.null(filters_list) && !all(purrr::map_vec(filters_list, \(x) is.function(x)))){
-    cli::cli_abort("If provided, filters_list must be a list of fishery filter functions. One or more list items is not a function.")
+    fram_abort("If provided, filters_list must be a list of fishery filter functions. One or more list items is not a function.")
   }
 
   validate_flag(warn)
@@ -339,10 +339,10 @@ plot_stock_mortality_time_step <- function(fram_db,
   }
 
   if(!is.null(filters_list) && !is.list(filters_list)){
-    cli::cli_abort("If provided, filters_list must be a list of fishery filter functions.")
+    fram_abort("If provided, filters_list must be a list of fishery filter functions.")
   }
   if(!is.null(filters_list) && !all(purrr::map_vec(filters_list, \(x) is.function(x)))){
-    cli::cli_abort("If provided, filters_list must be a list of fishery filter functions. One or more list items is not a function.")
+    fram_abort("If provided, filters_list must be a list of fishery filter functions. One or more list items is not a function.")
   }
 
   species_used = fetch_table_(fram_db, "RunID") |>